Д/з Модуляция(сделал)
